create README and add Page Views
The README file is the default page to be displayed on this repo, it's really helpful when you complete the necessary details here.

The page views is another one. 

When we add this service, it will count every hit of this repo. And this will guide us more about the visitors each day and will indicate the total views. For me, this is very helpful both for us and for those will view this repo seeing this page views. If there is the website built from this repo, it can be simply added there too.